requirements.txt
annotated-types==0.7.0
anyio==4.11.0
cached-property==2.0.1
certifi==2025.8.3
click==8.3.0
cloudinary==1.44.1
Deprecated==1.2.18
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.118.0
h11==0.16.0
httptools==0.6.4
idna==3.10
inflection==0.5.1
limits==5.6.0
mypy_extensions==1.1.0
packaging==25.0
psycopg2-binary==2.9.10
pydantic==2.11.10
pydantic_core==2.33.2
PyJWT==2.10.1
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.3
redis==6.4.0
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
SQLAlchemy==2.0.43
sqlalchemy-orm==1.2.10
starlette==0.48.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
wrapt==1.17.3
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
